Howto for implement parsedmarc
As DMARc is a great way to monitor your anti spam and anti phishing
setup but you need external tools to monitor and visualize the results
we show how to setup your own toolchain to achieve this task.

We show how to configure elasticsearch, kibana and parsedmarc to
implement a free solution for dmarc monitoring and visualization.