Change link format (#121)
To be more compliant with default jekyll environments we change the link
format form `YYYY-MM-DD-<title` to `YYYY/MM/DD/<title>`.
As there are many old posts with links in the old format out in the
wilde we add `jekyll-redirect-from` plugin to rewrite old links to new link style

To achieve this there were some internal links to be fixed too. All old
post got property `redirect_from` to handle the necessary rewrites.